test: simplify require_env tests and drop unused hint parameter
